requirements.txt
# Core ML (XTTS compatible)
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2

# Transformers (XTTS compatible)
transformers==4.36.2

# Speech-to-Text
faster-whisper==0.10.0

# Embeddings + RAG
sentence-transformers==2.2.2
huggingface_hub==0.25.2
faiss-cpu==1.7.4

# LLM runtime
llama-cpp-python==0.2.72

# Voice Cloning
TTS==0.22.0

# Audio utilities (compatible with numpy 1.22.0)
librosa==0.9.2
soundfile==0.12.1

# Utilities
tqdm
requests
